+++ engine/src/main/java/com/etherblood/luna/engine/Circle.java
@@ -1,9 +1,9 @@
 package com.etherblood.luna.engine;
 
 public record Circle(
-        int x,
-        int y,
-        int radius
+        long x,
+        long y,
+        long radius
 ) {
     public Circle {
         if (radius <= 0) {
@@ -15,20 +15,20 @@ public record Circle(
         return translate(position.x(), position.y());
     }
 
-    public Circle translate(int x, int y) {
+    public Circle translate(long x, long y) {
         return new Circle(this.x + x, this.y + y, radius);
     }
 
     public boolean contains(Vector2 pos) {
-        int dX = pos.x() - x;
-        int dY = pos.y() - y;
+        long dX = pos.x() - x;
+        long dY = pos.y() - y;
         return dX * dX + dY * dY < radius * radius;
     }
 
     public boolean intersects(Circle other) {
-        int dX = other.x() - x;
-        int dY = other.y() - y;
-        int radiusSum = other.radius() + radius;
+        long dX = other.x() - x;
+        long dY = other.y() - y;
+        long radiusSum = other.radius() + radius;
         return dX * dX + dY * dY < radiusSum * radiusSum;
     }
 }
