+++ engine/src/main/java/com/etherblood/luna/engine/Damagebox.java
@@ -2,6 +2,6 @@ package com.etherblood.luna.engine;
 
 public record Damagebox(
         Circle shape,
-        int milliDamage
+        long milliDamage
 ) {
 }
